# docs/swagger/swagger.yaml
basePath: /
definitions:
...
  helpers.Response:
    properties:
      data: {}
      error: {}
      status:
        type: integer
      success:
        type: boolean
    type: object
  locations.RequestCreateLocationLevel:
    properties:
      name:
        type: string
      need_parent_id:
        type: boolean
      status:
        type: integer
    type: object
  locations.RequestCreateLocations:
    properties:
      code:
        type: string
      level_id:
        type: integer
      name:
        type: string
      parent_id:
        type: integer
      status:
        type: integer
    type: object
  locations.RequestUpdateLocationLevel:
    properties:
      name:
        type: string
      need_parent_id:
        type: boolean
      status:
        type: integer
    type: object
  locations.RequestUpdateLocations:
    properties:
      code:
        type: string
      level_id:
        type: integer
      name:
        type: string
      parent_id:
        type: integer
      status:
        type: integer
    type: object
...
info:
  contact: {}
  description: Go API service
  title: Go API
  version: "1.0"
paths:
...
  /locations/:
    get:
      parameters:
      - description: Code
        in: query
        name: code
        type: string
      - description: Name
        in: query
        name: name
        type: string
      - description: asc|desc
        in: query
        name: order
        type: string
      - description: Order by (default name)
        in: query
        name: order_by
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/helpers.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helpers.Response'
      summary: List locations
      tags:
      - locations
  /locations/{locationID}:
    delete:
      parameters:
      - description: Location ID
        in: path
        name: locationID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/helpers.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helpers.Response'
      security:
      - BearerAuth: []
      summary: Delete location
      tags:
      - locations
    get:
      parameters:
      - description: Location ID
        in: path
        name: locationID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/helpers.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helpers.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/helpers.Response'
      summary: Get location detail
      tags:
      - locations
    put:
      consumes:
      - application/json
      parameters:
      - description: Location ID
        in: path
        name: locationID
        required: true
        type: string
      - description: Request body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/locations.RequestUpdateLocations'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/helpers.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helpers.Response'
      security:
      - BearerAuth: []
      summary: Update location
      tags:
      - locations
  /locations/{locationID}/status/{status}:
    put:
      parameters:
      - description: Location ID
        in: path
        name: locationID
        required: true
        type: string
      - description: Status
        in: path
        name: status
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/helpers.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helpers.Response'
      security:
      - BearerAuth: []
      summary: Update location status
      tags:
      - locations
  /locations/create:
    post:
      consumes:
      - application/json
      parameters:
      - description: Request body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/locations.RequestCreateLocations'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/helpers.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helpers.Response'
      security:
      - BearerAuth: []
      summary: Create location
      tags:
      - locations
  /locations/levels/:
    get:
      parameters:
      - description: Name
        in: query
        name: name
        type: string
      - description: asc|desc
        in: query
        name: order
        type: string
      - description: Order by (default name)
        in: query
        name: order_by
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/helpers.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helpers.Response'
      security:
      - BearerAuth: []
      summary: List location levels
      tags:
      - location-levels
  /locations/levels/{levelID}:
    get:
      parameters:
      - description: Level ID
        in: path
        name: levelID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/helpers.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helpers.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/helpers.Response'
      security:
      - BearerAuth: []
      summary: Get location level detail
      tags:
      - location-levels
    put:
      consumes:
      - application/json
      parameters:
      - description: Level ID
        in: path
        name: levelID
        required: true
        type: string
      - description: Request body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/locations.RequestUpdateLocationLevel'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/helpers.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helpers.Response'
      security:
      - BearerAuth: []
      summary: Update location level
      tags:
      - location-levels
  /locations/levels/{levelID}/delete:
    delete:
      parameters:
      - description: Level ID
        in: path
        name: levelID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/helpers.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helpers.Response'
      security:
      - BearerAuth: []
      summary: Delete location level
      tags:
      - location-levels
  /locations/levels/{levelID}/status/{status}:
    put:
      parameters:
      - description: Level ID
        in: path
        name: levelID
        required: true
        type: string
      - description: Status
        in: path
        name: status
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/helpers.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helpers.Response'
      security:
      - BearerAuth: []
      summary: Update location level status
      tags:
      - location-levels
  /locations/levels/create:
    post:
      consumes:
      - application/json
      parameters:
      - description: Request body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/locations.RequestCreateLocationLevel'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/helpers.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helpers.Response'
      security:
      - BearerAuth: []
      summary: Create location level
      tags:
      - location-levels
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by the access token
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"